requirements.txt
# 核心依赖
torch>=1.9.0
torchvision>=0.10.0
torch-geometric>=2.0.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0

# 化学信息学
rdkit>=2021.03.0
deepchem>=2.5.0

# 数据处理
pytorch-lightning>=1.5.0
dgl>=0.7.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
networkx>=2.6.0

# 工具类
tqdm>=4.62.0
pyyaml>=5.4.0
joblib>=1.0.0
pillow>=8.3.0

# Jupyter支持
jupyter>=1.0.0
ipywidgets>=7.6.0

# 测试
pytest>=6.2.0
pytest-cov>=2.12.0